% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_comp.R
\name{c.any}
\alias{c.any}
\title{Compute the concentration of the source salts for specific elements inside a fermentationbroth}
\usage{
\method{c}{any}(ds, ele = "P", org = "Bacteria")
}
\arguments{
\item{ds}{dry substance mass in g/l}

\item{ele}{element abreveation as string can be list('P', 'N', 'K','Mg')}

\item{org}{organism as string can be list('Bacteria', 'Fungus')}
}
\value{
concentration of salt within the broth
}
\description{
Compute the concentration of the source salts for specific elements inside a fermentationbroth
}
\examples{

# c.any(ds = 50, ele = 'P', org = 'Fungus')
# c.any(ds = 50, ele = "N", org = 'Fungus')
}
