% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellVit.R
\name{CellVit}
\alias{CellVit}
\title{Compute the cell vitality}
\usage{
CellVit(unstained, stained)
}
\arguments{
\item{unstained}{cell count of unstained cells}

\item{stained}{cell count of stained cells}
}
\value{
vitality
}
\description{
Compute the cell vitality
}
\examples{
us <- c(55, 65, 49, 57)
st <- c(11, 12, 9, 6)

CellVit(unstained = us, stained = st)

}
