% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_4PL.R
\name{conc.4PL}
\alias{conc.4PL}
\title{Compute the concentration of an competitive ELISA based on the 4-Parameter-logistic function}
\usage{
conc.4PL(std.conc, std.resp, mes.resp)
}
\arguments{
\item{std.conc}{the concentration of the calibrators}

\item{std.resp}{the measured absorbance of the calibrators}

\item{mes.resp}{the measured absorbance of the samples of unknown concentration}
}
\value{
the model-fit and the concentration of the samples
}
\description{
Compute the concentration of an competitive ELISA based on the 4-Parameter-logistic function
}
