% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absorption_spectroscopy.R, R/conc_from_abs.R
\name{conc.from.abs}
\alias{conc.from.abs}
\title{Functions to compute concentration values in absorption-spectroscopy}
\usage{
conc.from.abs(abs, epsilon, cuvette = 1, dfac = 1)

conc.from.abs(abs, epsilon, cuvette = 1, dfac = 1)
}
\arguments{
\item{abs}{measured absorption}

\item{epsilon}{molar attenuation coefficient (without default)}

\item{cuvette}{cuvette factor, default is 1}

\item{dfac}{dilution factor, default is  1}
}
\value{
The concentration
}
\description{
\code{sum} This function possesses the ability to return the concentration of an unknown sample by comparing its absorption to a linear regression model
}
