% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilfact.R
\name{conc.per.dil}
\alias{conc.per.dil}
\title{Compute the concentration after a dilution step}
\usage{
conc.per.dil(vol.pre, vol.post, conc.pre)
}
\arguments{
\item{vol.pre}{volume in advance of the dilution step}

\item{vol.post}{volume after the dilution step}

\item{conc.pre}{concentration in advance of the dilution}
}
\value{
current concentration of the dilution after the dilution step
}
\description{
Compute the concentration after a dilution step
}
