% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_vis_96.R
\name{concVis96.trimmed}
\alias{concVis96.trimmed}
\title{Visualisation of a 96-Wellplate (concentration) including trimming of the first col and row}
\usage{
concVis96.trimmed(plate)
}
\arguments{
\item{plate}{the plate to be visualised as dataframe}
}
\value{
a neat looking plot with the absorption intensity as color-gradient
}
\description{
This method trims of the first first row and col of by default
}
\examples{

\dontrun{concVis96.trimmed(HSA1)}

}
