% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response_plot.R
\name{dose_response_plot}
\alias{dose_response_plot}
\title{A dose~response plot with a summary}
\usage{
dose_response_plot(conc, resp, xlab = "conc", ylab = "resp")
}
\arguments{
\item{conc}{the concentrations}

\item{resp}{the response of the system for conc}

\item{xlab}{lable for the abscissa}

\item{ylab}{lable for the ordinate}
}
\value{
a plot
}
\description{
A dose~response plot with a summary
}
