% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrodynamik_FATL.R
\name{eG.empir}
\alias{eG.empir}
\title{Empiric prediction function for the Gas holdup (eG)}
\usage{
eG.empir(uG)
}
\arguments{
\item{uG}{superficial velocity}
}
\value{
gas holdup prediction
}
\description{
Empiric prediction function for the Gas holdup (eG)
}
