% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrodynamik_FATL.R
\name{eG}
\alias{eG}
\title{Compute gas content  
Keep an eye on the units}
\usage{
eG(VG, VL)
}
\arguments{
\item{VG}{volume of gas}

\item{VL}{volume of liquid}
}
\value{
eG or the gas content
}
\description{
Compute gas content  
Keep an eye on the units
}
\examples{
eG( VG = c(1, 2, 3), VL = c(6, 12, 24) )
}
