% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kLa.R
\name{eGR_CMC}
\alias{eGR_CMC}
\title{Gas content upstream column for non nowton fluids}
\usage{
eGR_CMC(K2 = 0.465, a2 = 0.65, b2 = -1.06, c2 = -0.103, eta.eff, uGR, AD, AR)
}
\arguments{
\item{K2}{constant of proportionality}

\item{a2}{specific coefficient}

\item{b2}{specific coefficient}

\item{c2}{specific coefficient}

\item{eta.eff}{dynamic viscosity}

\item{uGR}{superficial velocity (riser)}

\item{AD}{surface area downer}

\item{AR}{surface area riser}
}
\value{
eGR gas content upstream column for non nowton fluids
}
\description{
Gas content upstream column for non nowton fluids
}
