% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrodynamik_FATL.R
\name{eGT}
\alias{eGT}
\title{compute the total gas content (eGT) of an airlift-reactor with a single riser and a single downer}
\usage{
eGT(eGR, eGD, VDD, VDR)
}
\arguments{
\item{eGR}{gas content of the riser}

\item{eGD}{gas content of the downer}

\item{VDD}{dispersion volume of the downer}

\item{VDR}{dispersion volume of the riser}
}
\value{
total gas content (eGT)
}
\description{
compute the total gas content (eGT) of an airlift-reactor with a single riser and a single downer
}
