% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kLa.R
\name{eta.eff}
\alias{eta.eff}
\title{rheological modeling for highly non-newtonian fluids estimate of Ostwald & de Waele}
\usage{
eta.eff(Kc = (0.3^0.7), uGR, m)
}
\arguments{
\item{Kc}{consistency index}

\item{uGR}{superficial velocity (riser)}

\item{m}{index}
}
\value{
dynamic viscosity
}
\description{
rheological modeling for highly non-newtonian fluids estimate of Ostwald & de Waele
}
