% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSA1.R
\docType{data}
\name{HSA1}
\alias{HSA1}
\title{A 95-Well-Plate on which a sandwich ELISA assay detecting human serum albumin  was performed, the plates name is Clara}
\format{
A tibble: 8 x 14:
\describe{
  \item{clara}{axis of letters}
  \item{s1}{blank}
  \item{s2}{std1} 
  \item{s2}{std2}
  \item{s3}{std3}
  \item{s3-s12}{Measurments}
\item{s13}{means std}
}
}
\usage{
HSA1
}
\description{
for a exact pipetting-scheme browse the original Biotech-vignette
}
\keyword{datasets}
