% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrodynamik_FATL.R
\name{kLa.eG}
\alias{kLa.eG}
\title{compute kLa via eG (gas holdup)}
\usage{
kLa.eG(eG, K = 0.61, p = 0.83)
}
\arguments{
\item{eG}{gas holdup (eG)}

\item{K}{a specific factor, default is 0.61}

\item{p}{K a specific factor, default is 0.81}
}
\value{
the kLa via eG
}
\description{
compute kLa via eG (gas holdup)
}
