% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kLa.R
\name{kLaD.korr.H2O}
\alias{kLaD.korr.H2O}
\title{Correlation model for the kLa in H2O}
\usage{
kLaD.korr.H2O(K3 = 0.076, AD, AR, a3 = -2, uGR, b3 = 0.8)
}
\arguments{
\item{K3}{constant of proportionality}

\item{AD}{surface area downer}

\item{AR}{surface area riser}

\item{a3}{specific coefficient}

\item{uGR}{superficial velocity (riser)}

\item{b3}{specific coefficient}
}
\description{
Correlation model for the kLa in H2O
}
