% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kLa.R
\name{kLaD.korr.nN}
\alias{kLaD.korr.nN}
\title{Non-newton-correlation model for the kLa}
\usage{
kLaD.korr.nN(
  K4 = (0.5 * (10^-2)),
  uGR,
  a4 = 0.65,
  DGL = (2.7 * 10^(-9)),
  b4 = 0.5,
  rho.liq = 1020,
  c4 = 1.03,
  AD,
  AR,
  d4 = 0.85,
  eta.eff,
  e4,
  sigma.liq = 0.071,
  f4
)
}
\arguments{
\item{K4}{constant of proportionality}

\item{uGR}{superficial velocity (riser)}

\item{a4}{specific coefficient}

\item{DGL}{diffusion coefficient of the gas within the liquid phase}

\item{b4}{specific coefficient}

\item{rho.liq}{density of the liquid}

\item{c4}{specific coefficient}

\item{AD}{surface area downer}

\item{AR}{surface area riser}

\item{d4}{specific coefficient}

\item{eta.eff}{dynamic viscosity}

\item{e4}{specific coefficient}

\item{sigma.liq}{surface tension of the liqud}

\item{f4}{specific coefficient}
}
\description{
Non-newton-correlation model for the kLa
}
