% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lineweaver_Burk.R
\name{Lineweaver_Burk}
\alias{Lineweaver_Burk}
\title{Plot a Lineweaver-Burk diagram and compute the ordinate intercept}
\usage{
Lineweaver_Burk(
  sub,
  vel,
  title = "Lineweaver-Burk-Plot",
  xlab = "1/sub",
  ylab = "1/vel"
)
}
\arguments{
\item{sub}{substrate concentration}

\item{vel}{enzyme velocity}

\item{title}{title of the plot}

\item{xlab}{lable of the abscissa}

\item{ylab}{lable of the ordinate}
}
\description{
Plot a Lineweaver-Burk diagram and compute the ordinate intercept
}
