% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{LinMod}
\alias{LinMod}
\title{Linear model for concentration evaluation}
\usage{
LinMod(abs, conc)
}
\arguments{
\item{abs}{list}

\item{conc}{list}
}
\value{
summary of the regression
}
\description{
This must be performed in advanvce of usage of the function conc_eval to produce a linear model suiteable
for concentration evaluation.
}
