% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kLa.R
\name{mix.t.H2O}
\alias{mix.t.H2O}
\title{Mixing time of H2O based on correlation}
\usage{
mix.t.H2O(K5, uGR, m5)
}
\arguments{
\item{K5}{constant of proportionality}

\item{uGR}{superficial velocity (riser)}

\item{m5}{specific coefficient}
}
\value{
mixing time of H2O based on correlation
}
\description{
Mixing time of H2O based on correlation
}
