% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kLa.R
\name{mix.t.nN}
\alias{mix.t.nN}
\title{Mixing time of non newtonian fluids}
\usage{
mix.t.nN(
  K6 = 571,
  DR,
  a6 = -0.5,
  AD,
  AR,
  b6 = -0.16,
  h.disp.Riser,
  c6 = -1.44,
  uGR,
  d6 = -0.46,
  eta.eff,
  e6 = 0.56
)
}
\arguments{
\item{K6}{constant of proportionality}

\item{DR}{diamter of the riser}

\item{a6}{specific coefficient}

\item{AD}{surface area downer}

\item{AR}{surface area riser}

\item{b6}{specific coefficient}

\item{h.disp.Riser}{height of the liquid (riser) gased}

\item{c6}{specific coefficient}

\item{uGR}{superficial velocity (riser)}

\item{d6}{specific coefficient}

\item{eta.eff}{dynamic viscosity}

\item{e6}{specific coefficient}
}
\description{
Mixing time of non newtonian fluids
}
