% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrodynamik_FATL.R
\name{mt}
\alias{mt}
\title{compute the mixing time}
\usage{
mt(H, D, uG, gforce = 9.81)
}
\arguments{
\item{H}{height of the reactor}

\item{D}{diameter of the reactor}

\item{uG}{superficial velocity}

\item{gforce}{g-force, default is 9.81}
}
\value{
the mixing time
}
\description{
compute the mixing time
}
