% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellVit.R
\name{MultiRate}
\alias{MultiRate}
\title{Compute the cell vitality after frosting}
\usage{
MultiRate(unstained_defro, unstained_24h)
}
\arguments{
\item{unstained_defro}{cell count of unstained (vital) cells immediately after defrosting}

\item{unstained_24h}{count of unstained (vital) cells 24 hours after defrosting}
}
\value{
multiplication count after frosting
}
\description{
Compute the cell vitality after frosting
}
\examples{
ud  <- c(12, 22, 15, 17)
u24 <-  c(9, 3, 5, 4)

MultiRate(unstained_defro =ud, unstained_24h = u24)

}
