% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGE_gen.R
\name{PAGE.gen}
\alias{PAGE.gen}
\title{Function to analyse a PAGE}
\usage{
PAGE.gen(marker.w, marker.d, unknown.d, plot.PAGE = FALSE)
}
\arguments{
\item{marker.w}{weight of marker protein}

\item{marker.d}{marker-band distance}

\item{unknown.d}{distance of a unknown sample-band}

\item{plot.PAGE}{want a plot? Set TRUE or FALSE}
}
\value{
approx protein-weight
}
\description{
Function to analyse a PAGE
}
