% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kLa.R
\name{PGVL}
\alias{PGVL}
\title{Power input of a bubble column reactor}
\usage{
PGVL(
  uGR,
  A.Riser = 0.00528,
  p,
  rho.liq = 997.048,
  gforce = 9.81,
  rho.gas = 1.2041,
  d.gaser = (10^-3),
  d.upcol,
  h.undisp.Riser,
  h.undisp.Downer,
  A.Downer = 0.00196,
  N = 52
)
}
\arguments{
\item{uGR}{superficial velocity (riser)}

\item{A.Riser}{column cross-section  (riser)}

\item{p}{room pressure}

\item{rho.liq}{density of the liquid}

\item{gforce}{g force}

\item{rho.gas}{density of the gas}

\item{d.gaser}{surface area of the gas nozzle}

\item{d.upcol}{surface area of the riser}

\item{h.undisp.Riser}{height of the liquid (riser) without gas}

\item{h.undisp.Downer}{height of the liquid (downer) without gas}

\item{A.Downer}{column cross-section (downer)}

\item{N}{count of gas outlets}
}
\value{
power input
}
\description{
Power input of a bubble column reactor
}
