% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MM_direkt_1.R
\name{plot_MM_direct}
\alias{plot_MM_direct}
\title{Create a non-linear Regression for a Michaelis-Menten type Enzyme with a little knowlege of its Km and Vmax}
\usage{
plot_MM_direct(
  sub,
  velo,
  title = "title",
  xlab = "abscisaa",
  ylab = "ordinate"
)
}
\arguments{
\item{sub}{substrate concentration}

\item{velo}{velocity}

\item{title}{title of the plot}

\item{xlab}{lable of the abscissa}

\item{ylab}{lable of the ordinate}
}
\value{
a plot
}
\description{
Create a non-linear Regression for a Michaelis-Menten type Enzyme with a little knowlege of its Km and Vmax
}
