% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regression.R
\name{plot_regression}
\alias{plot_regression}
\title{Generic plotting function for usage with any absorption spectroscopic data}
\usage{
plot_regression(abs, conc)
}
\arguments{
\item{abs}{list}

\item{conc}{list}
}
\description{
create a graph from obtained data of calibration standards
}
