% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zlokarnik.R
\name{PM}
\alias{PM}
\title{The power-input via the motor}
\usage{
PM(P.perc, rpm, F1 = 0.107, F2 = 0.062, FC = 0.753)
}
\arguments{
\item{P.perc}{motor power in percent}

\item{rpm}{rotation per minute}

\item{F1}{armature voltage / rpm}

\item{F2}{armature current / motor power in percent}

\item{FC}{degree of efficiency}
}
\value{
power-input
}
\description{
The power-input via the motor
}
