% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilfact.R
\name{samp.vol}
\alias{samp.vol}
\title{Compute sample (Vorlage) volumes 
V_vor = (V_{soll} * c_{soll}) /  c_{ist}}
\usage{
samp.vol(conc.aim, vol.aim = 1, conc.is)
}
\arguments{
\item{conc.aim}{aim concentration (C_soll)}

\item{vol.aim}{volume that is aimed (V_soll)}

\item{conc.is}{current concentration (C_ist)}
}
\value{
sample volumes
}
\description{
Compute sample (Vorlage) volumes 
V_vor = (V_{soll} * c_{soll}) /  c_{ist}
}
