% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbsSpktr.R
\docType{data}
\name{Spektren}
\alias{Spektren}
\title{Absorbance measurements}
\format{
A tibble: 110 x 3
}
\usage{
Spektren
}
\description{
This dataset was arbitrary chosen for demonstrations purposes
}
\keyword{datasets}
