% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbsSpktr.R
\docType{data}
\name{Spektrum}
\alias{Spektrum}
\title{Absorption measurement}
\format{
A tibble: 391 x 2
}
\usage{
Spektrum
}
\description{
This dataset was arbitrary chosen for demonstrations purposes
}
\keyword{datasets}
