% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STR.Prog.Assay.R
\docType{data}
\name{STR.Prog.Assay}
\alias{STR.Prog.Assay}
\title{A calibrationcurve to test 4-parameter-linear-models}
\format{
A tibble: 8 x 14:
\describe{
  \item{conc}{Concentration of each measure}
  \item{E1}{std1} 
  \item{E2}{std2}
  \item{E3}{std3}
\item{MW}{means std}
}
}
\usage{
STR.Prog.Assay
}
\description{
for more information browse the original Biotech-vignette
}
\keyword{datasets}
