% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrodynamik_FATL.R
\name{uG}
\alias{uG}
\title{compute the superficial velocity}
\usage{
uG(velG, A)
}
\arguments{
\item{velG}{gas velocity}

\item{A}{flow surface}
}
\value{
the superficial velocity
}
\description{
compute the superficial velocity
}
