% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrodynamik_FATL.R
\name{VDi}
\alias{VDi}
\title{Compute the dispersion volume
mind the units}
\usage{
VDi(Ai, HDi)
}
\arguments{
\item{Ai}{column cross-section}

\item{HDi}{gasified height}
}
\value{
the dispersion volume
}
\description{
Compute the dispersion volume
mind the units
}
\examples{
VDi(Ai = 0.45, HDi = 1.33)
}
