% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellVit.R
\name{vitality}
\alias{vitality}
\title{Vitality of a cell population}
\usage{
vitality(stained, unstained)
}
\arguments{
\item{stained}{cell count of stained cells}

\item{unstained}{cell count of unstained cells}
}
\value{
Vitality of the cell population
}
\description{
Vitality of a cell population
}
\examples{
vitality(stained = 3, unstained = 77)
}
