% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellVit.R
\name{vitrate}
\alias{vitrate}
\title{Vitality rate of a cell population}
\usage{
vitrate(unstained_prefro, unstained_defro)
}
\arguments{
\item{unstained_prefro}{cell count of unstained cells prior to frosting}

\item{unstained_defro}{cell count of unstained cells after frosting}
}
\value{
Vitality rate of a cell population
}
\description{
Vitality rate of a cell population
}
\examples{
vitrate(unstained_prefro = 33, unstained_defro = 12)
}
