% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzyme_activity.R
\name{vol.act}
\alias{vol.act}
\title{generic functions to determin protein activity}
\usage{
vol.act(vol.tot = 1, dfac = 1, delta.E, vol.test, epsilon, cuvette = 1)
}
\arguments{
\item{vol.tot}{total volume, default is 1}

\item{dfac}{dilution factor, default is 1}

\item{delta.E}{Delta E (without default)}

\item{vol.test}{test volume (without default)}

\item{epsilon}{molar attenuation coefficient (without default)}

\item{cuvette}{cuvette factor or path length, default is 1}
}
\description{
All Volumes should be given with same metrics
}
