% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellVit.R
\name{yield}
\alias{yield}
\title{Compute yield count}
\usage{
yield(unstained_24h, unstained_prefro)
}
\arguments{
\item{unstained_24h}{count of unstained (vital) cells 24 hours after defrosting}

\item{unstained_prefro}{cell count of unstained (vital) cells in advance of frosting}
}
\value{
yield count
}
\description{
Compute yield count
}
\examples{
uP  <- c(88, 73, 72, 97)
u24 <-  c(9, 3, 5, 4)
yield(unstained_prefro = uP, unstained_24h = u24)
}
