% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_eval.R
\name{conc_eval}
\alias{conc_eval}
\title{concentration evaluation}
\usage{
conc_eval(abs_P, abs_std, conc_std)
}
\arguments{
\item{abs_P}{absorption of a unknown sample}

\item{abs_std}{absorption of calibrationstandards}

\item{conc_std}{concentration of calibrationstandards}
}
\value{
float
}
\description{
concentration evaluation
}
