% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisy.R
\docType{data}
\name{daisy}
\alias{daisy}
\title{Different values measured during a fermentation of Escherichia coli in a bioreactor named daisy}
\format{

}
\usage{
daisy
}
\description{
Different values measured during a fermentation of Escherichia coli in a bioreactor named daisy
}
\keyword{datasets}
