% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kLa.R
\name{eGR_H2O}
\alias{eGR_H2O}
\title{gas content of upstream column for a bubble column reactor}
\usage{
eGR_H2O(K1, uGr, m1)
}
\arguments{
\item{K1}{specific coefficient}

\item{uGr}{superficial velocity (riser)}

\item{m1}{specific coefficient}
}
\value{
gas content upstream column for water
}
\description{
gas content of upstream column for a bubble column reactor
}
