% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HSA1}
\alias{HSA1}
\title{Whatever HSA1 is.}
\format{
A data frame with 8 rows and 14 variables:
\describe{
   \item{clara}{whatever this is}
   \item{s1}{something else, up to s 12}
   \item{Std}{no idea}
}
}
\usage{
HSA1
}
\description{
A dataset with xyz...
}
\keyword{datasets}
